- add isAdmin check on post edit and delete

+++ client/src/api/PostsRequests.js
@@ -13,5 +13,5 @@ API.interceptors.request.use((req) => {
 
 export const getTimelinePosts= (id)=> API.get(`/posts/${id}/timeline`);
 export const likePost=(id, userId)=> API.put(`posts/${id}/like`, {userId: userId})
-export const deletePost=(id, userId)=> API.delete(`posts/${id}`,  { data: { userId: userId }})
+export const deletePost=(id, userId, isAdmin)=> API.delete(`posts/${id}`,  { data: { userId: userId, isAdmin: isAdmin}})
 export const updatePost=(id, data)=> API.put(`posts/${id}`,  {data: data})
